services/requirements.txt
anyio==4.7.0
asgiref==3.8.1
certifi==2024.12.14
Django==5.1
django-admin-interface==0.28.9
django-colorfield==0.11.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
semantic-version==2.10.0
setuptools==74.0.0
sniffio==1.3.1
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
updatable==0.7.1
uritemplate==4.1.1
